// src/components/stapi_client/typescript/types/strapi_generated/contentTypes.d.ts
export interface ApiCaseStudyCaseStudy extends Schema.CollectionType {
  collectionName: 'case_studies';
  info: {
    singularName: 'case-study';
    pluralName: 'case-studies';
    displayName: 'CaseStudy';
    description: '';
  };
  options: {
    draftAndPublish: true;
  };
  pluginOptions: {
    i18n: {
      localized: true;
    };
  };
  attributes: {
    title: Attribute.String &
      Attribute.Required &
      Attribute.SetPluginOptions<{
        i18n: {
          localized: true;
        };
      }> &
      Attribute.DefaultTo<'Domy\u015Blny tytu\u0142'>;
    subtitle: Attribute.Text &
      Attribute.Required &
      Attribute.SetPluginOptions<{
        i18n: {
          localized: true;
        };
      }> &
      Attribute.DefaultTo<'Domy\u015Blna tre\u015B\u0107'>;
    image: Attribute.Media &
      Attribute.Required &
      Attribute.SetPluginOptions<{
        i18n: {
          localized: false;
        };
      }>;
    authors: Attribute.Relation<
      'api::case-study.case-study',
      'oneToMany',
      'api::author.author'
    >;
    slug: Attribute.UID<'api::case-study.case-study', 'title'> &
      Attribute.Required;
    date: Attribute.Date &
      Attribute.Required &
      Attribute.SetPluginOptions<{
        i18n: {
          localized: false;
        };
      }>;
    body: Attribute.DynamicZone<
      [
        'layout.editor',
        'layout.embed-video',
        'layout.fancy-link',
        'layout.note',
        'layout.quote',
        'layout.simple-image'
      ]
    > &
      Attribute.Required &
      Attribute.SetPluginOptions<{
        i18n: {
          localized: true;
        };
      }>;
    createdAt: Attribute.DateTime;
    updatedAt: Attribute.DateTime;
    publishedAt: Attribute.DateTime;
    createdBy: Attribute.Relation<
      'api::case-study.case-study',
      'oneToOne',
      'admin::user'
    > &
      Attribute.Private;
    updatedBy: Attribute.Relation<
      'api::case-study.case-study',
      'oneToOne',
      'admin::user'
    > &
      Attribute.Private;
    localizations: Attribute.Relation<
      'api::case-study.case-study',
      'oneToMany',
      'api::case-study.case-study'
    >;
    locale: Attribute.String;
  };
}